fix(ui): make buttonVariants server-safe to fix brands page crash
The brands page is a Server Component but called buttonVariants() (for the
Product Tour link added in #225). buttonVariants was exported from button.tsx,
which is a 'use client' module, so the export became a client reference and
threw at render: 'Attempted to call buttonVariants() from the server but
buttonVariants is on the client.'

Move the cva definition into its own non-client module
(components/ui/button-variants.ts) and import buttonVariants into the brands
page from there. button.tsx re-exports it unchanged, so the 6 existing client
callers keep working with no change.
